Only set CPUSchedulingPolicy once
As of af05e97 both CPUSchedulingPolicy and IOSchedulingClass were
disabled in the freeswitch systemd service for compatibility with LXD.
But however, the first option was then re-enabled in the config.yml,
triggering a restart of freeswitch every time.

But outside of virtualised environments, there is on need to remove the
CPUSchedulingPolicy option[^1]. And the IOSchedulingClass is set to
`realtime` by default[^2], but was also disabled for the LXD
compatibility.

To stop freeswitch from being restarted when there is no actual
configuration change and dial back the removal of IOSchedulingClass the
tasks where consolidated into one task so it's idempotent.

The default value for `bbb_freeswitch_ioschedule_realtime` is set to
`false` by default for backwards compatibility with setups that require
this configuration.

Close #59

[^1]: https://docs.bigbluebutton.org/2.2/troubleshooting.html#freeswitch-fails-to-start-with-a-setscheduler-error
[^2]: https://github.com/signalwire/freeswitch/blob/15ad4c23e259c1c2dff58f89e9124e5f36dd2e94/debian/freeswitch-systemd.freeswitch.service#L30