Implement bbb_api_demos_enable variable
Currently the variable `bbb_api_demos_enable` is listed in the README,
however it is not used in the playbook. Practically, the `bbb-demo`
package is not installed and the API demo is not working.

This PR adds support for the installation and, quite importantly,
removal of the `bbb-demo` apt package.