convert unix LF to dos CRLF in logback.xml
Although I would prefer LF over CRLF, upstream does not so for
logback.xml. After an upgrade of bbb the ansible-role overwrites the
whole file, instead of just changing the loglevels.
It is hard to see the actual diffrences made to the file.
So we should follow upstream here.
This refers to the older PR ebbba-org/ansible-role-bigbluebutton/pull/215