fix: Allow dial-in for breakout rooms
The current default dialplan config would only accept 5 digits for a conference PIN, but breakout rooms have longer PINs. When trying to dial into a breakout room, the user is thrown into the main-room instead, which may cause some confusion.

The new default allows 5 to 7 digits, which is enough for up to 99 breakout rooms per meeting. The timeout between playback of conf-pin.wav and the first digit is increased from 7 to 10 seconds to give users more time to look up the PIN. The maximum time between two digits was reduced from 7 to 5 seconds to help users that do not know that short PINs need to be terminated with a '#'.