Improved networking configuration
You can now define different IPs for services to bind to, and for services to advertise as their public IP. This is important in NAT settings where the real public IP is not actually assigned to any of the network interfaces.

Freeswitch now listens on locaslhost for the control socket and sip (tunneled through nginx, search for sipjsHackViaWs) and binds top the correct IP instead of 0.0.0.0 for other services.